+++ apps/www/src/ui/SearchResults/index.tsx
@@ -90,19 +90,25 @@ const SearchResults: FC<Props> = ({ maxResults = 32 }) => {
                             </a>
                             .
                         </p>
-                        <div className={styles.results}>
-                            {nodeResults.map(node => (
-                                <IllustratedNodeView key={node.uuid} value={node} short />
-                            ))}
-                            {resolutions.map(resolution => (
-                                <IllustratedNodeView
-                                    key={resolution.uuid}
-                                    caption={<ExternalResolutionCaption value={resolution} />}
-                                    value={resolution.node}
-                                    short
-                                />
-                            ))}
-                        </div>
+                        {nodeResults.length > 0 && (
+                            <div className={styles.results}>
+                                {nodeResults.map(node => (
+                                    <IllustratedNodeView key={node.uuid} value={node} short />
+                                ))}
+                            </div>
+                        )}
+                        {resolutions.length > 0 && (
+                            <div className={styles.results}>
+                                {resolutions.map(resolution => (
+                                    <IllustratedNodeView
+                                        key={resolution.uuid}
+                                        caption={<ExternalResolutionCaption value={resolution} />}
+                                        value={resolution.node}
+                                        short
+                                    />
+                                ))}
+                            </div>
+                        )}
                     </Fragment>
                 )}
             </Container>