+++ apps/www/CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+-   Upgraded `next` to `13.3.1`.
+
 ### Deprecated
 
 ### Fixed